+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.39.0] - 2026-04-16
+### Added
+- Added `info_message` and `info_code` fields to `DependencyResponse.Dependencies` message in Dependencies responses as the definitive fields for reporting the outcome of processing each component
+### Removed
+- **Breaking change:** Removed `error_message` and `error_code` fields from `DependencyResponse.Dependencies` message in Dependencies responses. Clients must migrate to `info_message`/`info_code`
+
 ## [0.38.0] - 2026-04-16
 ### Added
 - Added `info_message` and `info_code` fields to `ComponentLocationInfo`, `ComponentLocation`, `ContributorResponse.Purls`, and `OriginResponse.Purls` messages in Geoprovenance responses
@@ -287,6 +293,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Vulnerabilities
 - Added REST endpoint support for each service also
 
+[0.39.0]: https://github.com/scanoss/papi/compare/v0.38.0...v0.39.0
 [0.38.0]: https://github.com/scanoss/papi/compare/v0.37.0...v0.38.0
 [0.37.0]: https://github.com/scanoss/papi/compare/v0.36.0...v0.37.0
 [0.36.0]: https://github.com/scanoss/papi/compare/v0.35.0...v0.36.0